+++ backend/services/migration-service-backend/src/main/resources/db/migration/postgres/auth/V202602220809300000__Create_table_User_Tenants.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS auth.user_tenant (
+    user_id UUID REFERENCES auth.user(id),
+    tenant_id UUID REFERENCES auth.tenant(id),
+    PRIMARY KEY (user_id, tenant_id)
+);