+++ backend/services/migration-service-backend/src/main/resources/db/migration/postgres/auth/V202602220813240000__Create_table_Resources.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS auth.resources (
+    id UUID PRIMARY KEY,
+    code VARCHAR(100) UNIQUE NOT NULL,
+    description TEXT
+);