+++ backend/services/migration-service-backend/CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ## Alterações não lançadas
 
+Fev 22, 2026
+- [migration-service-backend-0001] - Criação das tabelas para o serviço de autenticão.
+
 Jan 18, 2026
 - [master] - Configurações iniciais do projeto.